fix: bail early when installing channels on non-ubuntu
Fixes #11615

+++ packages/playwright-core/bin/reinstall_msedge_stable_linux.sh
@@ -8,6 +8,17 @@ if [[ $(arch) == "aarch64" ]]; then
   exit 1
 fi
 
+if [[ ! -f "/etc/os-release" ]]; then
+  echo "ERROR: cannot install on unknown linux distribution"
+  exit 1
+fi
+
+ID=$(bash -c 'source /etc/os-release && echo $ID')
+if [[ "${ID}" != "ubuntu" ]]; then
+  echo "ERROR: cannot install on $ID distribution - only Ubuntu is supported"
+  exit 1
+fi
+
 # 1. make sure to remove old stable if any.
 if dpkg --get-selections | grep -q "^microsoft-edge-stable[[:space:]]*install$" >/dev/null; then
   apt-get remove -y microsoft-edge-stable
